Source code Deliverable 4
While creating this code I've learned how to create functions, how to pass parameters, how to use and initialize variables, and how to write comments on code. I also researched and installed different libraries for python itself.

In this commit to my code, I created a user interface that connected to an excel file using openpyxl this was to serve as my database for my usernames and passwords for the program in the future. I also created a way to create new usernames, validate the information entered in the text boxes for the username and password to sign you into the specific users profile.